Разнесла классы по отдельным файлам библиотек. Переписала импорты. Тесты запускаются.
