+++ hollidays/test_view.py
@@ -1,9 +1,12 @@
 from bok_choy.web_app_test import WebAppTest
-from hollidays.pages.login import LoginPage
-from hollidays.pages.orders import OrdersPage, CREATE
+
+from hollidays.pages.elements import CREATE
+from hollidays.pages.pages import LoginPage
+from hollidays.pages.pages import OrdersPage
 from hollidays.pages.views import View
 from hollidays import BaseWebTest
 import datetime
+
 
 
 class TestView(BaseWebTest):
@@ -43,8 +46,23 @@ class TestView(BaseWebTest):
     #     assert view_page.modal_is_visible()
     #     assert view_page.modal.input_visible()
 
+    # def test_create_new_view(self):
+    #     """
+    #     Тест создает новое представление.
+    #     1. залогинеться в ситемему
+    #     2. Создать новое представление
+    #     3. проверить что пользователь находится в новом представлении
+    #     # ! Важно: пока не реализовано удаление созданного представления, удалять руками
+    #     """
+    #     new_view_name = "New_view " + (datetime.datetime.now()).strftime("%m-%d-%H:%M:%S")
+    #     # self.login('lara@lara.ru', '123123')
+    #     view_page = View(self.browser)
+    #     view_page._create_view(new_view_name)
+    #     assert view_page._get_current_view_name() == new_view_name
+
     def test_create_new_view(self):
         """
+        Делаю новый тест на основе новой архитектуре классов
         Тест создает новое представление.
         1. залогинеться в ситемему
         2. Создать новое представление
@@ -52,10 +70,13 @@ class TestView(BaseWebTest):
         # ! Важно: пока не реализовано удаление созданного представления, удалять руками
         """
         new_view_name = "New_view " + (datetime.datetime.now()).strftime("%m-%d-%H:%M:%S")
-        # self.login('lara@lara.ru', '123123')
-        view_page = View(self.browser)
-        view_page._create_view(new_view_name)
-        assert view_page._get_current_view_name() == new_view_name
+        view_page = OrdersPage(self.browser)
+        assert view_page.is_browser_on_page()
+        view_page.dropdown.select(CREATE)  # выбираем Create
+        assert view_page.modal_is_visible()
+
+
+        #assert view_page._get_current_view_name() == new_view_name
 
     def test_detete_new_view(self):
         """
